separate unit tests from integration tests and edit the makefiles accordingly
