+++ test/Makefile
@@ -1,11 +1,16 @@
-.PHONY: test
+.PHONY: testint
 
-test:
+testint:
 	@LINODE_FIXTURE_MODE="play" \
 	LINODE_TOKEN="awesometokenawesometokenawesometoken" \
 	LINODE_API_VERSION="v4beta" \
 	GO111MODULE="on" \
 	go test -v ./integration $(ARGS)
+
+.PHONY: testunit
+
+testunit:
+	go test -v ./unit/...
 
 .PHONY: smoketest
