+++ test/integration/lke_clusters_test.go
@@ -6,7 +6,6 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/jarcoal/httpmock"
 	"github.com/linode/linodego"
 	k8scondition "github.com/linode/linodego/k8s/pkg/condition"
 )
@@ -216,32 +215,6 @@ func TestLKEVersion_GetMissing(t *testing.T) {
 
 	if e.Code != 404 {
 		t.Errorf("should have received a 404 Code requesting a missing version, got %v", e.Code)
-	}
-}
-
-func TestLKECluster_Regenerate(t *testing.T) {
-	client := createMockClient(t)
-
-	requestData := linodego.LKEClusterRegenerateOptions{
-		KubeConfig:   true,
-		ServiceToken: false,
-	}
-
-	httpmock.RegisterRegexpResponder("POST", mockRequestURL(t, "clusters/1234/regenerate"),
-		mockRequestBodyValidate(t, requestData, nil))
-
-	if _, err := client.RegenerateLKECluster(context.Background(), 1234, requestData); err != nil {
-		t.Fatal(err)
-	}
-}
-
-func TestLKECluster_DeleteServiceToken(t *testing.T) {
-	client := createMockClient(t)
-
-	httpmock.RegisterRegexpResponder("DELETE", mockRequestURL(t, "clusters/1234/servicetoken"), httpmock.NewStringResponder(200, "{}"))
-
-	if err := client.DeleteLKEClusterServiceToken(context.Background(), 1234); err != nil {
-		t.Fatal(err)
 	}
 }
