+++ test/integration/util_test.go
@@ -1,14 +1,9 @@
 package integration
 
 import (
-	"regexp"
 	"strconv"
 	"testing"
 	"time"
-
-	"github.com/linode/linodego/internal/testutil"
-
-	"github.com/jarcoal/httpmock"
 )
 
 func assertDateSet(t *testing.T, compared *time.Time) {
@@ -16,14 +11,6 @@ func assertDateSet(t *testing.T, compared *time.Time) {
 	if compared == nil || *compared == emptyTime {
 		t.Errorf("Expected date to be set, got %v", compared)
 	}
-}
-
-func mockRequestBodyValidate(t *testing.T, expected interface{}, response interface{}) httpmock.Responder {
-	return testutil.MockRequestBodyValidate(t, expected, response)
-}
-
-func mockRequestURL(t *testing.T, path string) *regexp.Regexp {
-	return testutil.MockRequestURL(path)
 }
 
 func assertSliceContains[T comparable](t *testing.T, slice []T, target T) {
