+++ test/unit/instance_stats_test.go
@@ -1,4 +1,4 @@
-package integration
+package unit
 
 import (
 	"context"
