+++ test/unit/object_storage_test.go
@@ -1,4 +1,4 @@
-package integration
+package unit
 
 import (
 	"context"
@@ -6,16 +6,6 @@ import (
 
 	"github.com/jarcoal/httpmock"
 )
-
-func TestObjectStorage_Get_Transfer(t *testing.T) {
-	client, teardown := createTestClient(t, "fixtures/TestObjectStorage_transfer")
-	defer teardown()
-
-	_, err := client.GetObjectStorageTransfer(context.Background())
-	if err != nil {
-		t.Errorf("unable to get object storage transfer data : %s", err)
-	}
-}
 
 func TestObjectStorage_Cancel(t *testing.T) {
 	client := createMockClient(t)
