+++ test/unit/profile_security_questions_test.go
@@ -1,4 +1,4 @@
-package integration
+package unit
 
 import (
 	"context"
